remove hardcoded package names from rtest.sh
Fix flake8 to not recurse into built directories
(in an ideal world this would be based on .gitignore)

Fix mypy to not crash on setuptools missing type stubs

Add .mypy.ini

And now let me rant for a moment about this naming.
`setup.cfg` is right out because I don't need to do distutils stuff
and none of the linting is related to it anyways.
I want to have each tools' settings in a separate file.
`tox.ini` is out additionally because I don't need or want tox.
Now, flake8 only supports `.flake8`,
but mypy supports `.mypy` or `mypy`, but with the ini extension.
WHY CAN'T YOU BE CONSISTENT!? 😭
(I can't wait for something to break
because I used the unicode instead of the ASCII shortcode.)